replace hwmon, cdrom and hal with udisks for media detection
Remove the old cdrom code to poll a drive and merge the remaining
hwmon and hal into one module. HAL itself is replaced by DeviceKit
(UDisks) and therefore, the hwmon changed its name to udisks.

The code was tested with two USB sicks and one DVD. It should work
as the old code did because some parts were just copied. But be
aware that it may be buggy and that it won't detect all devices
correctly.