Add functions to scan a directory without monitor
The scan() client function allws the client to scan a
directory and all subdirectories similar to the monitor()
function. The directories are only scanned and NOT monitored
and therefore, changes will only be visible on the next
manual scan. This is usefull if the cient knows about some
directories that do not change often.